V1.0

+++ src/com/example/phonetopc/model/FileLoadInfo.java
@@ -3,7 +3,7 @@ package com.example.phonetopc.model;
 public class FileLoadInfo {
 	private String file_name;
 	private String file_type;
-	private int file_size;
+	private double file_size;
 
 	public String getFile_name() {
 		return file_name;
@@ -21,11 +21,11 @@ public class FileLoadInfo {
 		this.file_type = file_type;
 	}
 
-	public int getFile_size() {
+	public double getFile_size() {
 		return file_size;
 	}
 
-	public void setFile_size(int file_size) {
+	public void setFile_size(double file_size) {
 		this.file_size = file_size;
 	}
 
@@ -35,7 +35,7 @@ public class FileLoadInfo {
 		this.file_type = "";
 	}
 
-	public FileLoadInfo(String arg0, String arg1, int arg2) {
+	public FileLoadInfo(String arg0, String arg1, double arg2) {
 		this.file_name = arg0;
 		this.file_type = arg1;
 		this.file_size = arg2;
